# utils.match: Speed up by ~45% by replacing `suppress(KeyError)` with `in`