@has2k1 2018-06-21
Thanks for the profiling, `utils.match` is a bottleneck. I'll look more into it.